#990 Split the writer round-trip omnibus along its seams
WriterRoundTripTest said it wrote a flat REQUIRED INT32 file and asserted the values
survived. It had become 2197 lines and 83 cases spanning batch validation, nulls,
page and row-group boundaries, CRC, dictionary selection, compression, statistics
and every nested shape -- the default home for anything without one, while every
other class in the package is scoped to a single concern.

The seams were already visible in the method names, so this follows them: contract,
layout, dictionary, compression, statistics and nested each get the class their tests
were already grouped under, and the original keeps the values-and-nulls story its
name promises. The schemas and readers more than one of them needs move to
WriterTestSupport rather than being duplicated, which is how two test classes come to
disagree about what a file said.

No case is added, removed or changed: 83 before, 83 after.
